+++ lib/packages/wikibase/data-model/DataModel/Entity/PropertyId.php
@@ -3,7 +3,6 @@
 namespace Wikibase\DataModel\Entity;
 
 use InvalidArgumentException;
-use Wikibase\Property;
 
 /**
  * @since 0.5