+++ lib/packages/wikibase/data-model/DataModel/References.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wikibase;
+namespace Wikibase\DataModel;
 
 /**
  * List of Reference objects.