+++ lib/packages/wikibase/data-model/DataModel/Snak/SnakRole.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wikibase;
+namespace Wikibase\DataModel\Snak;
 
 /**
  * Enum with snak roles.